+++ src/cli.js
@@ -1,5 +1,5 @@
 const yargs = require('yargs');
-const createTwilioFunction = require('./create-twilio-function');
+const { handler, describe, cliInfo } = require('./command');
 
 function cli(cwd) {
   yargs.help();
@@ -9,41 +9,18 @@ function cli(cwd) {
 
   yargs.default('path', cwd);
 
-  yargs.usage('Creates a new Twilio Function project');
+  yargs.usage(describe);
   yargs.command(
     '$0 <name>',
-    'Creates a new Twilio Function project',
+    describe,
     command => {
       command.positional('name', {
         describe: 'Name of your project.',
         type: 'string'
       });
-      command.options({
-        'account-sid': {
-          alias: 'a',
-          describe: 'The Account SID for your Twilio account',
-          type: 'string'
-        },
-        'auth-token': {
-          alias: 't',
-          describe: 'Your Twilio account Auth Token',
-          type: 'string'
-        },
-        'skip-credentials': {
-          describe:
-            "Don't ask for Twilio account credentials or import them from the environment",
-          type: 'boolean',
-          default: false
-        },
-        'import-credentials': {
-          describe:
-            'Import credentials from the environment variables TWILIO_ACCOUNT_SID and TWILIO_AUTH_TOKEN',
-          type: 'boolean',
-          default: false
-        }
-      });
+      command.options(cliInfo.options);
     },
-    argv => createTwilioFunction(argv)
+    argv => handler(argv)
   );
 
   return yargs;